fix beanstalk actions

+++ src/routes/gitwebhook.js
@@ -5,12 +5,17 @@ export default function gitwebhook(req, res) {
     const hmac = crypto.createHmac('sha1', process.env.SECRET);
     const sig = `sha1=${hmac.update(JSON.stringify(req.body)).digest('hex')}`;
     if (
+        req.headers['request url'].includes('glitch.me') &&
         req.headers['x-github-event'] === 'pull_request' &&
         crypto.timingSafeEqual(Buffer.from(sig), Buffer.from(req.headers['x-hub-signature']))
     ) {
         res.sendStatus(200);
-        const { state, merged } = req.body.pull_request;
-        if (state === 'closed' && merged) {
+        const {
+            state,
+            merged,
+            base: { ref },
+        } = req.body.pull_request;
+        if (state === 'closed' && merged && ref === 'master') {
             const commands = [
                 'git fetch origin master',
                 'git reset --hard origin/master',
